Extract into RevisionXmlBuilder
