+++ psalm.xml
@@ -6,6 +6,7 @@
     xmlns="https://getpsalm.org/schema/config"
     xsi:schemaLocation="https://getpsalm.org/schema/config ../../vendor/vimeo/psalm/config.xsd"
     errorBaseline="psalm-baseline.xml"
+	autoloader="../../includes/AutoLoader.php"
 >
     <projectFiles>
         <directory name="includes" />
@@ -17,5 +18,6 @@
 	<issueHandlers>
 		<MissingPropertyType errorLevel="suppress" />
 		<MissingClosureReturnType errorLevel="suppress" />
+		<UndefinedConstant errorLevel="suppress" />
 	</issueHandlers>
 </psalm>
